[Refactor] #110 - 코스 그리고 업로드 API를 CourseRouter로 위치 변경
